% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{comp_scat2d}
\alias{comp_scat2d}
\title{Compare two sets of 2D points by plotting a scatterplot, optionally rotating the data such that they are as similar as possible (useful when visualizing projections which are invariante to unitary transformations (rotations and reflections). A will be in black, B will be in red.}
\usage{
comp_scat2d(A, B, rot = FALSE, scale_ax = FALSE, obnoxious = FALSE)
}
\arguments{
\item{A}{A first set of points, should be of the same dimensions as B}

\item{B}{A first set of points, should be of the same dimensions as A}

\item{rot}{If TRUE, solves the orthogonal procrustes problem to rotate B to match A.}

\item{scale_ax}{If TRUE, scales each axis so as to minimize error (performed after procrustes analysis if rot is also TRUE).}
}
\description{
Compare two sets of 2D points by plotting a scatterplot, optionally rotating the data such that they are as similar as possible (useful when visualizing projections which are invariante to unitary transformations (rotations and reflections). A will be in black, B will be in red.
}
