% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_max.R
\name{em_plsvn}
\alias{em_plsvn}
\title{EM posterior maximization for PLSV.}
\usage{
em_plsvn(docs, K, V, P, eta, gamma, beta, make_plot = FALSE,
  THETA_init = NULL, PSI_init = NULL, PHI_init = NULL,
  THETA_fix = list(), PSI_fix = list(), verbose = FALSE, thresh = 0.01,
  max_iters = 1000)
}
\arguments{
\item{docs}{A term frequency matrix, that is, one with a row for each document, a column for each vocab word, and integer entries indicating the occurence of a word in a doc.}

\item{K}{The number of topics, an integer scalar.}

\item{V}{The number of unique words, an integer scalar.}

\item{P}{The dimensionality of the embedding space, an integer, usually 2.}

\item{eta}{The exchangible dirichlet prior on words in a topic.}

\item{beta}{The precision for topic locations, a positive scalar.}

\item{make_plot}{A boolean, if TRUE, will make a ggplot visualization of the topics and documents, with topics in red.}

\item{THETA_init}{Either a real matrix with as many rows as docs has and P many columns, giving an initial value for THETA, or a scalar character, either 'smart' or 'random'. See details..}

\item{PSI_init}{Either a real matrix with K many rows and P many columns, giving an initial value for PSI, or a scalar character, either 'smart' or 'random'. See details.}

\item{PHI_init}{Either a matrix of K many V-1-simplex valued rows, giving the initial value for PHI, or a scalar character, either 'smart' or 'random'. See details.}

\item{THETA_fix}{A list of lists, used to fix rows of THETA to a given value. Each sublist has two elements: 'ind' and 'val'. 'ind' Indicates the row, 1-index, of THETA to fix, and 'val', a real valued P-vector, indicates the value to fix it to.}

\item{verbose}{Boolean, if TRUE, tells you what the biggest jump of on screen coords is.}

\item{thresh}{The threshold for the entire EM algo; if the biggest absolute difference between coordinates onscreen is less than this, the algo stops.}

\item{max_iters}{The maximum number of EM iterations allowed, an integer scalar.}

\item{gama}{The precision for document locations, a positive scalar.}
}
\value{
A list containing ests, a list with PHI, the topic by document matrix, THETA, the document locations in P-D space, and PSI, the topic locations in P-D space.
}
\description{
Perform the EM algorithm described in Iwata et al to estimate a PSLV model.
}
