% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_max.R
\name{g_nilp}
\alias{g_nilp}
\title{The Gradient of the Incomplete Log Posterior}
\usage{
g_nilp(PHI, THETA, PSI, docs, eta, gamma, beta, soft_PHI)
}
\description{
this boii accepts a list of docs for right now.
}
