% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_max.R
\name{g_nlip}
\alias{g_nlip}
\title{The Gradient of the Log Incomplete Posterior}
\usage{
g_nlip(PHI, THETA, PSI, docs, eta, gamma, beta, soft_PHI)
}
\description{
this boii accepts a list of docs for right now.
}
