% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{gen_plsv}
\alias{gen_plsv}
\title{Generate from the PLSV model}
\usage{
gen_plsv(K, V, M, N.mu, P, eta, gamma, beta)
}
\arguments{
\item{K}{The number of topics, an integer scalar.}

\item{V}{The number of unique words, an integer scalar.}

\item{M}{The number of documents, an integer scalar.}

\item{N.mu}{The average number of words per document.}

\item{P}{The dimensionality of the embedding space, an integer, usually 2.}

\item{eta}{The exchangible dirichlet prior on words in a topic.}

\item{beta}{The precision for topic locations, a positive scalar.}

\item{gama}{The precision for document locations, a positive scalar.}
}
\value{
A list containing the generated data, in particular, PHI is the topic by word matrix (rows sum to 1), PSI is a topic location matrix, docs is a list of generated documents, each entry denotes index of vocab, THETA is the document locations, and Z is a list of matrices, with a 1 denoting which topic the word belonged to.
}
\description{
Generate from the PLSV model
}
