% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_max.R
\name{nclp}
\alias{nclp}
\title{Negative Complete Log Posterior}
\usage{
nclp(PHI, THETA, PSI, docs, eta, gamma, beta)
}
\arguments{
\item{THETA}{M by P real valued matrix, giving P-d locations of documents.}

\item{PSI}{K by P real valued matrix, giving P-d locations of topics}

\item{docs}{A term frequency matrix, that is, one with a row for each document, a column for each vocab word, and integer entries indicating the occurence of a word in a doc.}

\item{eta}{The exchangible dirichlet prior on words in a topic.}

\item{beta}{The precision for topic locations, a positive scalar.}

\item{PHI_n}{A K by V matrix with simplex valued rows giving the the probabilty of words (cols) in topics (rows).}

\item{gama}{The precision for document locations, a positive scalar.}

\item{make_plot}{A boolean, if TRUE, will make a ggplot visualization of the topics and documents, with topics in red.}
}
\value{
A scalar, giving the negative log posterior density at the point.
}
\description{
The log posterior with the latent class assignments integrated out, evaluated at PHI_n, THETA, PSI as specified, given docs.
}
