% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_max.R
\name{nlip}
\alias{nlip}
\title{Negative Log Incomplete Posterior}
\usage{
nlip(PHI, THETA, PSI, docs, eta, gamma, beta, soft_PHI)
}
\arguments{
\item{PHI}{A K by V matrix with simplex valued rows giving the the probabilty of words (cols) in topics (rows).}

\item{THETA}{M by P real valued matrix, giving P-d locations of documents.}

\item{PSI}{K by P real valued matrix, giving P-d locations of topics}

\item{docs}{A term frequency matrix, that is, one with a row for each document, a column for each vocab word, and integer entries indicating the occurence of a word in a doc.}

\item{eta}{The exchangible dirichlet prior on words in a topic.}

\item{beta}{The precision for topic locations, a positive scalar.}

\item{soft_PHI}{A boolean, if TRUE, PHI was provided on the logodds scales.}

\item{gama}{The precision for document locations, a positive scalar.}
}
\value{
A scalar, giving the negative log posterior density at the point.
}
\description{
The log posterior with the latent class assignments integrated out, evaluated at PHI, THETA, PSI as specified, given docs.
}
