% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_max.R
\name{num_post_plsv}
\alias{num_post_plsv}
\title{Numerically Maximize complete Posterior.}
\usage{
num_post_plsv(docs, K, V, P, eta, gamma, beta, make_plot = FALSE,
  THETA_fix = list(), PSI_fix = list())
}
\arguments{
\item{docs}{A term frequency matrix, that is, one with a row for each document, a column for each vocab word, and integer entries indicating the occurence of a word in a doc.}

\item{K}{The number of topics, an integer scalar.}

\item{V}{The number of unique words, an integer scalar.}

\item{P}{The dimensionality of the embedding space, an integer, usually 2.}

\item{eta}{The exchangible dirichlet prior on words in a topic.}

\item{beta}{The precision for topic locations, a positive scalar.}

\item{make_plot}{A boolean, if TRUE, will make a ggplot visualization of the topics and documents, with topics in red.}

\item{THETA_fix}{A list of lists, used to fix rows of THETA to a given value. Each sublist has two elements: 'ind' and 'val'. 'ind' Indicates the row, 1-index, of THETA to fix, and 'val', a real valued P-vector, indicates the value to fix it to.}

\item{gama}{The precision for document locations, a positive scalar.}
}
\value{
A list containing ests, a list with PHI, the topic by document matrix, THETA, the document locations in P-D space, and PSI, the topic locations in P-D space.
}
\description{
Numerically maximize the complete posterior of the PLSV model.
}
