% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{orth_proc}
\alias{orth_proc}
\title{Solve the orthogonal procrustes problem}
\usage{
orth_proc(A, B, comp_stress = TRUE)
}
\arguments{
\item{A}{A Matrix of the same dimension as B}

\item{B}{A Matrix of the same dimension of A of which a rotation will be computed which puts it closest to A.}

\item{comp_stress}{Should we compute and return the F norm of A - WB, where W is the rotation?}
}
\value{
If comp_stress, a list with elements W, the rotation matrix, and stress, the scalar norm of the  difference, or just the matrix W if not.
}
\description{
Solve the orthogonal procrustes problem
}
